requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langgraph==0.0.40
faiss-cpu==1.7.4
groq==0.28.0
streamlit==1.29.0
requests==2.31.0
pypdf==3.17.0
python-dotenv==1.0.0
prometheus-client==0.19.0
faker==22.0.0
pandas==2.0.0
arize-phoenix==4.0.0
opentelemetry-sdk==1.24.0